fix: test fixed for chest and new voting Power calculation. BoosterTimestamp instead of creationTimestamp in VestingPosition

+++ test/invariant/chest/utils/BaseSetup.sol
@@ -25,7 +25,7 @@ contract ChestHarness is Chest {
 
     function exposed_calculateBooster(
         ChestHarness.VestingPosition memory vestingPosition,
-        uint256 timestamp
+        uint48 timestamp
     ) external view returns (uint120) {
         return calculateBooster(vestingPosition, timestamp);
     }
